+++ test/OpenTelemetry.OpAmp.Client.Tests/WsTransportTest.cs
@@ -2,34 +2,43 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using OpenTelemetry.OpAmp.Client.Internal;
-using OpenTelemetry.OpAmp.Client.Internal.Transport.Http;
+using OpenTelemetry.OpAmp.Client.Internal.Transport.WebSocket;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using OpenTelemetry.OpAmp.Client.Tests.Tools;
 using Xunit;
 
 namespace OpenTelemetry.OpAmp.Client.Tests;
 
-public class PlainHttpTransportTest
+public class WsTransportTest
 {
     [Theory]
     [InlineData(true)]
     [InlineData(false)]
-    public async Task PlainHttpTransport_SendReceiveCommunication(bool useSmallPackets)
+    public async Task WsTransport_SendReceiveCommunication(bool useSmallPackets)
     {
         // Arrange
-        using var opAmpServer = new OpAmpFakeHttpServer(useSmallPackets);
+        using var opAmpServer = new OpAmpFakeWebSocketServer(useSmallPackets);
         var opAmpEndpoint = opAmpServer.Endpoint;
 
         using var mockListener = new MockListener();
         var frameProcessor = new FrameProcessor();
         frameProcessor.Subscribe(mockListener);
 
-        var httpTransport = new PlainHttpTransport(opAmpEndpoint, frameProcessor);
+        using var cts = new CancellationTokenSource();
+        using var wsTransport = new WsTransport(opAmpEndpoint, frameProcessor);
+        await wsTransport.StartAsync(cts.Token);
 
-        var mockFrame = FrameGenerator.GenerateMockAgentFrame(useSmallPackets);
+        // Send only small packets, currently sending large package is not supported in WsTransport
+        var mockFrame = FrameGenerator.GenerateMockAgentFrame(useSmallPackets: true);
 
         // Act
-        await httpTransport.SendAsync(mockFrame.Frame, CancellationToken.None);
+        await wsTransport.SendAsync(mockFrame.Frame, cts.Token);
+
+        mockListener.WaitForMessages(TimeSpan.FromSeconds(30));
+
+        await wsTransport.StopAsync(cts.Token);
+
+        cts.Cancel();
 
         // Assert
         var serverReceivedFrames = opAmpServer.GetFrames();