+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 * Add `db.query.parameter.<key>` attribute(s) to query spans if opted into using
-  the `SetDbQueryParameters` option. Not supported on .NET Framework.
-  ([#3015](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3015))
+  the `OTEL_DOTNET_EXPERIMENTAL_SQLCLIENT_ENABLE_TRACE_DB_QUERY_PARAMETERS`
+  environment variable. Not supported on .NET Framework.
+  ([#3015](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3015),
+  [#3081](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3081))
 
 * Fix activities not being stopped on .NET Framework when using a global activity
   listener.